Split the lock-audit test by concern
One file held three kinds of check with different epistemics; each now has a home whose name predicts its content:

- tests/supply-chain-audit.test.mjs (rename): the structural invariants proved from committed artifacts -- locks, IOCs, install-script inventory, npmrc keys, git-dep pins, install-chain pins, workflow audit
- tests/runner-lint.test.mjs: the sanctioned-runner-forms text scan, whose lint-not-boundary header now labels a whole file; the deny-pattern self-scan exclusion covers both definition files
- tests/test-wiring.test.mjs: the suite-wiring guard (test:repo args resolve), which was never a supply-chain check -- and the natural home for the planned reporter-based test-count floor

Pure move: zero assertion changes, same 146-test count; the wiring guard itself proved red mid-migration on the stale test:repo name. File references in scripts, docs, and the changelog updated.
