+++ package.json
@@ -70,7 +70,7 @@
     "set:version": "npm run -s _spv --",
     "test:full": "npm run _test:full:pre && npm run check && npm run test:website && npm run _test:full:common",
     "test:lychee": "node --test 'tests/lychee/**/*.test.mjs'",
-    "test:repo": "node --test tests/hugo-versions.test.mjs tests/lock-audit.test.mjs tests/mermaid-version.test.mjs 'scripts/**/*.test.mjs' 'theme/scripts/**/*.test.mjs' 'tests/fixture-site/**/*.test.mjs'",
+    "test:repo": "node --test tests/hugo-versions.test.mjs tests/mermaid-version.test.mjs tests/runner-lint.test.mjs tests/supply-chain-audit.test.mjs tests/test-wiring.test.mjs 'scripts/**/*.test.mjs' 'theme/scripts/**/*.test.mjs' 'tests/fixture-site/**/*.test.mjs'",
     "test:site-equivalence": "node tests/site-equivalence/full-vs-shared.mjs --check",
     "test:smoke": "node tests/smoke.test.mjs",
     "test:website": "npm run -C docsy.dev test",
